Changing the Ansatz Depths
Sorry that this folder is really messy. I tried to make multiple copies of my code to attempt to run them in parallel. Comparing the results to the ones in Seed Generator, it seems that the seed is also affecting the algorithm's output. e.g. For seed = 214, the state after the first iteration is void. The state after the first iter is extracted from the Parameters list, compiled by store_intermediate_result() function provided in the qiskit tutorial, records the state for each iteration of vqe. For certain seeds such as 214, the entire Parameters list is void, which is really interesting (strange).